Machine Learning Code
Python code that reads in the previous results from the season and predicts all match results based off how
each team has played at home and away. A Matrix is created and the numbers in that matrix are added up to
get an estimated percentage of each team winning. This process takes home team advantage and form throughout
the season into account in its predictions